// src/pages/Profile/types/types.ts
export type Movies = {
	id: string;
	title: string;
	description: string;
	rating: number;
	image: string;
};

export type Series = {
	id: string;
	title: string;
	description: string;
	rating: number;
	image: string;
};

export type Animes = {
	id: string;
	title: string;
	description: string;
	rating: number;
	image: string;
};